+++ Casks/tiny-clips.rb
@@ -1,7 +1,7 @@
 cask "tiny-clips" do
   auto_updates true
-  version "1.5.2"
-  sha256 "bbd777eb2a9ad38d07fc031c499172d5527eb3d8bd4a19d64d413c7dae64f578"
+  version "1.5.4"
+  sha256 "65d2cc681f66461853d0c676229d2727cae24def4f2a4ce8b05fbe32985b8242"
 
   url "https://github.com/jamesmontemagno/tiny-clips/releases/download/v#{version}-mac/TinyClips-v#{version}-mac.zip"
   name "TinyClips"